requirements.txt
# requirements.txt — Multi-Cloud Observability Collector
# ─────────────────────────────────────────────────────────────────────────────
# Core (always required)
PyYAML>=6.0

# ── AWS ───────────────────────────────────────────────────────────────────────
boto3>=1.34

# ── Azure ─────────────────────────────────────────────────────────────────────
azure-identity>=1.15
azure-mgmt-resource>=23.0
azure-mgmt-compute>=30.0
azure-mgmt-web>=7.0
azure-mgmt-sql>=3.0
azure-monitor-query>=1.3

# ── GCP ───────────────────────────────────────────────────────────────────────
google-cloud-monitoring>=2.18
google-cloud-logging>=3.8
google-cloud-compute>=1.14
google-auth>=2.27
requests>=2.31

# ── Storage backends ──────────────────────────────────────────────────────────
# SQLite is built into Python — no extra package needed.
# Uncomment the line below if using PostgreSQL or TimescaleDB:
# psycopg2-binary>=2.9

# ── Prometheus exporter ───────────────────────────────────────────────────────
prometheus-client>=0.19
